persist selected spark type
