+++ PluginsAndFeatures/azure-toolkit-for-intellij/resources/META-INF/plugin.xml
@@ -393,8 +393,8 @@
     </action>
 
     <action id="Actions.SubmitSparkApplicationAction"
-            class="com.microsoft.azure.hdinsight.spark.actions.SparkSubmitJobAction" text="Submit Spark Application to HDInsight"
-            description="Submit Spark Application to HDInsight" icon="/icons/Spark.png"/>
+            class="com.microsoft.azure.hdinsight.spark.actions.SparkSubmitJobAction" text="Submit Spark Application"
+            description="Submit Spark Application" icon="/icons/Spark.png"/>
 
     <action id="AzureToolkit.AzureSignIn" class="com.microsoft.azuretools.ijidea.actions.AzureSignInAction" text="Azure Sign In..." />
     <action id="AzureToolkit.SelectSubscriptions" class="com.microsoft.azuretools.ijidea.actions.SelectSubscriptionsAction" text="Select Subscriptions..."
@@ -495,20 +495,17 @@
       <reference ref="Spark.RunScalaLivyConsole"/>
     </group>
 
-    <group id="SetDefaultRunConfigGroup" text="Set default spark run configuration" description="Default Spark Run Configuration" popup="true">
-      <add-to-group group-id="RunMenu" anchor="after" relative-to-action="SparkJobDebug"/>
+    <group id="SetDefaultRunConfigGroup" text="Default Spark Application Type" description="Select default spark application type" popup="true">
+      <add-to-group group-id="ToolsMenu" anchor="last"/>
       <add-to-group group-id="RunContextPopupGroup" anchor="after" relative-to-action="SparkJobDebug"/>
-      <action id="SetDefaultSparkRunConfigurationTypeNone"
-        class="com.microsoft.azure.hdinsight.spark.run.action.SetDefaultSparkRunConfigurationTypeAction$None" text="None" >
-      </action>
       <action id="SetDefaultSparkRunConfigurationTypeHDInsight"
-              class="com.microsoft.azure.hdinsight.spark.run.action.SetDefaultSparkRunConfigurationTypeAction$HDInsight" text="HDInsight" >
+              class="com.microsoft.azure.hdinsight.spark.run.action.DefaultSparkApplicationTypeAction$HDInsight" text="HDInsight" >
       </action>
       <action id="SetDefaultSparkRunConfigurationTypeCosmosSpark"
-              class="com.microsoft.azure.hdinsight.spark.run.action.SetDefaultSparkRunConfigurationTypeAction$CosmosSpark" text="CosmosSpark" >
+              class="com.microsoft.azure.hdinsight.spark.run.action.DefaultSparkApplicationTypeAction$CosmosSpark" text="CosmosSpark" >
       </action>
       <action id="SetDefaultSparkRunConfigurationTypeCosmosServerlessSpark"
-              class="com.microsoft.azure.hdinsight.spark.run.action.SetDefaultSparkRunConfigurationTypeAction$CosmosServerlessSpark" text="CosmosServerlessSpark" >
+              class="com.microsoft.azure.hdinsight.spark.run.action.DefaultSparkApplicationTypeAction$CosmosServerlessSpark" text="CosmosServerlessSpark" >
       </action>
     </group>
   </actions>
